+++ src/search/pdbs/cegar.cc
@@ -4,36 +4,179 @@
 #include "utils.h"
 
 #include "../option_parser.h"
+#include "../task_proxy.h"
 
 #include "../task_utils/task_properties.h"
 
 #include "../utils/countdown_timer.h"
+#include "../utils/logging.h"
 #include "../utils/rng.h"
 
 #include <limits>
+#include <unordered_set>
 
 using namespace std;
 
 namespace pdbs {
+/*
+  This is used as a "collection entry" in the CEGAR algorithm. It stores
+  the PDB (and with that, the pattern) and an optimal plan (in the wildcard
+  format) for that PDB if it exists or unsolvable is true otherwise. It can
+  be marked as "solved" to ignore it in further iterations of the CEGAR
+  algorithm.
+*/
+class PatternInfo {
+    shared_ptr<PatternDatabase> pdb;
+    vector<vector<OperatorID>> plan;
+    bool unsolvable;
+    bool solved;
+
+public:
+    PatternInfo(
+        const shared_ptr<PatternDatabase> &&pdb,
+        const vector<vector<OperatorID>> &&plan,
+        bool unsolvable)
+        : pdb(move(pdb)),
+          plan(move(plan)),
+          unsolvable(unsolvable),
+          solved(false) {}
+
+    const shared_ptr<PatternDatabase> &get_pdb() const {
+        return pdb;
+    }
+
+    const Pattern &get_pattern() const {
+        return pdb->get_pattern();
+    }
+
+    const vector<vector<OperatorID>> &get_plan() const {
+        return plan;
+    }
+
+    bool is_unsolvable() const {
+        return unsolvable;
+    }
+
+    void mark_as_solved() {
+        solved = true;
+    }
+
+    bool is_solved() {
+        return solved;
+    }
+};
+
+struct Flaw {
+    int collection_index;
+    int variable;
+
+    Flaw(int collection_index, int variable)
+        : collection_index(collection_index),
+          variable(variable) {
+    }
+};
+
+using FlawList = vector<Flaw>;
+
+class CEGAR {
+    const int max_pdb_size;
+    const int max_collection_size;
+    const double max_time;
+    const bool use_wildcard_plans;
+    const utils::Verbosity verbosity;
+    shared_ptr<utils::RandomNumberGenerator> rng;
+    const shared_ptr<AbstractTask> &task;
+    const TaskProxy task_proxy;
+    const vector<FactPair> &goals;
+    unordered_set<int> blacklisted_variables;
+
+    vector<unique_ptr<PatternInfo>> pattern_collection;
+    /*
+      Map each variable of the task which is contained in the collection to the
+      collection index at which the pattern containing the variable is stored.
+    */
+    unordered_map<int, int> variable_to_collection_index;
+    int collection_size;
+
+    void print_collection() const;
+    bool time_limit_reached(const utils::CountdownTimer &timer) const;
+
+    unique_ptr<PatternInfo> compute_pattern_info(Pattern &&pattern) const;
+    void compute_initial_collection();
+
+    /*
+      Check if operator op is applicable in state, ignoring blacklisted
+      variables. If it is, return an empty (flaw) list. Otherwise, return the
+      violated preconditions.
+    */
+    FlawList get_violated_preconditions(
+        int collection_index,
+        const OperatorProxy &op,
+        const vector<int> &current_state) const;
+    /*
+      Try to apply the plan of the pattern at the given index in the
+      concrete task starting at the given state. During application,
+      blacklisted variables are ignored. If plan application succeeds,
+      return an empty flaw list. Otherwise, return all precondition variables
+      of all operators of the failing plan step. When the method returns,
+      current is the last state reached when executing the plan.
+     */
+    FlawList apply_plan(int collection_index, vector<int> &current_state) const;
+    /*
+      Use apply_plan to generate flaws. Return true if there are no flaws and
+      no blacklisted variables, in which case the concrete task is solved.
+      Return false in all other cases. Append new flaws to the passed-in flaws.
+    */
+    bool get_flaws_for_pattern(
+        int collection_index, const State &concrete_init, FlawList &flaws);
+    /*
+      Use get_flaws_for_pattern for all patterns of the collection. Append
+      new flaws to the passed-in flaws. If the task is solved by the plan of
+      any pattern, return the collection index of that pattern. Otherwise,
+      return -1.
+    */
+    int get_flaws(const State &concrete_init, FlawList &flaws);
+
+    // Methods related to refining.
+    void add_pattern_for_var(int var);
+    bool can_merge_patterns(int index1, int index2) const;
+    void merge_patterns(int index1, int index2);
+    bool can_add_variable_to_pattern(int index, int var) const;
+    void add_variable_to_pattern(int collection_index, int var);
+    void refine(const FlawList &flaws);
+public:
+    CEGAR(
+        int max_pdb_size,
+        int max_collection_size,
+        double max_time,
+        bool use_wildcard_plans,
+        utils::Verbosity verbosity,
+        const shared_ptr<utils::RandomNumberGenerator> &rng,
+        const shared_ptr<AbstractTask> &task,
+        const vector<FactPair> &goals,
+        unordered_set<int> &&blacklisted_variables = unordered_set<int>());
+    PatternCollectionInformation compute_pattern_collection();
+};
+
 CEGAR::CEGAR(
     int max_pdb_size,
     int max_collection_size,
-    bool use_wildcard_plans,
     double max_time,
+    bool use_wildcard_plans,
     utils::Verbosity verbosity,
     const shared_ptr<utils::RandomNumberGenerator> &rng,
     const shared_ptr<AbstractTask> &task,
-    vector<FactPair> &&goals,
+    const vector<FactPair> &goals,
     unordered_set<int> &&blacklisted_variables)
     : max_pdb_size(max_pdb_size),
       max_collection_size(max_collection_size),
-      use_wildcard_plans(use_wildcard_plans),
       max_time(max_time),
+      use_wildcard_plans(use_wildcard_plans),
       verbosity(verbosity),
       rng(rng),
       task(task),
       task_proxy(*task),
-      goals(move(goals)),
+      goals(goals),
       blacklisted_variables(move(blacklisted_variables)),
       collection_size(0) {
 #ifndef NDEBUG
@@ -51,43 +194,6 @@ CEGAR::CEGAR(
         }
     }
 #endif
-    if (verbosity >= utils::Verbosity::NORMAL) {
-        utils::g_log << "options of the CEGAR algorithm for computing a pattern collection: " << endl;
-        utils::g_log << "max pdb size: " << max_pdb_size << endl;
-        utils::g_log << "max collection size: " << max_collection_size << endl;
-        utils::g_log << "wildcard plans: " << use_wildcard_plans << endl;
-        utils::g_log << "Verbosity: ";
-        switch (verbosity) {
-        case utils::Verbosity::SILENT:
-            utils::g_log << "silent";
-            break;
-        case utils::Verbosity::NORMAL:
-            utils::g_log << "normal";
-            break;
-        case utils::Verbosity::VERBOSE:
-            utils::g_log << "verbose";
-            break;
-        case utils::Verbosity::DEBUG:
-            utils::g_log << "debug";
-            break;
-        }
-        utils::g_log << endl;
-        utils::g_log << "max time: " << max_time << endl;
-        utils::g_log << "goal variables: ";
-        for (const FactPair &goal : this->goals) {
-            utils::g_log << goal.var << ", ";
-        }
-        utils::g_log << endl;
-        utils::g_log << "blacklisted variables: ";
-        if (this->blacklisted_variables.empty()) {
-            utils::g_log << "none";
-        } else {
-            for (int var : this->blacklisted_variables) {
-                utils::g_log << var << ", ";
-            }
-        }
-        utils::g_log << endl;
-    }
 }
 
 void CEGAR::print_collection() const {
@@ -108,7 +214,7 @@ bool CEGAR::time_limit_reached(
     const utils::CountdownTimer &timer) const {
     if (timer.is_expired()) {
         if (verbosity >= utils::Verbosity::NORMAL) {
-            utils::g_log << "time limit reached." << endl;
+            utils::g_log << "CEGAR time limit reached" << endl;
         }
         return true;
     }
@@ -129,7 +235,7 @@ unique_ptr<PatternInfo> CEGAR::compute_pattern_info(Pattern &&pattern) const {
     if (pdb->get_value(initial_state.get_unpacked_values()) == numeric_limits<int>::max()) {
         unsolvable = true;
         if (verbosity >= utils::Verbosity::VERBOSE) {
-            utils::g_log << "Projection onto pattern " << pdb->get_pattern()
+            utils::g_log << "projection onto pattern " << pdb->get_pattern()
                          << " is unsolvable" << endl;
         }
     } else {
@@ -443,14 +549,52 @@ void CEGAR::refine(const FlawList &flaws) {
 
     if (!added_var) {
         if (verbosity >= utils::Verbosity::VERBOSE) {
-            utils::g_log << "could not add var/merge patterns due to size "
-                "limits. Blacklisting." << endl;
+            utils::g_log << "could not add var/merge pattern containing var "
+                         << "due to size limits, blacklisting var" << endl;
         }
         blacklisted_variables.insert(var);
     }
 }
 
 PatternCollectionInformation CEGAR::compute_pattern_collection() {
+    if (verbosity >= utils::Verbosity::NORMAL) {
+        utils::g_log << "CEGAR options:" << endl;
+        utils::g_log << "max pdb size: " << max_pdb_size << endl;
+        utils::g_log << "max collection size: " << max_collection_size << endl;
+        utils::g_log << "max time: " << max_time << endl;
+        utils::g_log << "wildcard plans: " << use_wildcard_plans << endl;
+        utils::g_log << "verbosity: ";
+        switch (verbosity) {
+        case utils::Verbosity::SILENT:
+            utils::g_log << "silent";
+            break;
+        case utils::Verbosity::NORMAL:
+            utils::g_log << "normal";
+            break;
+        case utils::Verbosity::VERBOSE:
+            utils::g_log << "verbose";
+            break;
+        case utils::Verbosity::DEBUG:
+            utils::g_log << "debug";
+            break;
+        }
+        utils::g_log << endl;
+        utils::g_log << "goal variables: ";
+        for (const FactPair &goal : this->goals) {
+            utils::g_log << goal.var << ", ";
+        }
+        utils::g_log << endl;
+        utils::g_log << "blacklisted variables: ";
+        if (this->blacklisted_variables.empty()) {
+            utils::g_log << "none";
+        } else {
+            for (int var : this->blacklisted_variables) {
+                utils::g_log << var << ", ";
+            }
+        }
+        utils::g_log << endl;
+    }
+
     utils::CountdownTimer timer(max_time);
     compute_initial_collection();
     int iteration = 1;
@@ -527,9 +671,67 @@ PatternCollectionInformation CEGAR::compute_pattern_collection() {
     return pattern_collection_information;
 }
 
-void add_implementation_notes_to_parser(options::OptionParser &parser) {
+PatternCollectionInformation generate_pattern_collection_with_cegar(
+    int max_pdb_size,
+    int max_collection_size,
+    double max_time,
+    bool use_wildcard_plans,
+    utils::Verbosity verbosity,
+    const shared_ptr<utils::RandomNumberGenerator> &rng,
+    const shared_ptr<AbstractTask> &task,
+    const vector<FactPair> &goals,
+    unordered_set<int> &&blacklisted_variables) {
+    CEGAR cegar(
+        max_pdb_size,
+        max_collection_size,
+        max_time,
+        use_wildcard_plans,
+        verbosity,
+        rng,
+        task,
+        goals,
+        move(blacklisted_variables));
+    return cegar.compute_pattern_collection();
+}
+
+PatternInformation generate_pattern_with_cegar(
+    int max_pdb_size,
+    double max_time,
+    bool use_wildcard_plans,
+    utils::Verbosity verbosity,
+    const shared_ptr<utils::RandomNumberGenerator> &rng,
+    const shared_ptr<AbstractTask> &task,
+    const FactPair &goal,
+    unordered_set<int> &&blacklisted_variables) {
+    vector<FactPair> goals = {goal};
+    CEGAR cegar(
+        max_pdb_size,
+        max_pdb_size,
+        max_time,
+        use_wildcard_plans,
+        verbosity,
+        rng,
+        task,
+        goals,
+        move(blacklisted_variables));
+    PatternCollectionInformation collection_info = cegar.compute_pattern_collection();
+    shared_ptr<PatternCollection> new_patterns = collection_info.get_patterns();
+    if (new_patterns->size() > 1) {
+        cerr << "CEGAR limited to one goal computed more than one pattern" << endl;
+        utils::exit_with(utils::ExitCode::SEARCH_CRITICAL_ERROR);
+    }
+
+    Pattern &pattern = new_patterns->front();
+    shared_ptr<PDBCollection> new_pdbs = collection_info.get_pdbs();
+    shared_ptr<PatternDatabase> &pdb = new_pdbs->front();
+    PatternInformation result(TaskProxy(*task), move(pattern));
+    result.set_pdb(pdb);
+    return result;
+}
+
+void add_cegar_implementation_notes_to_parser(options::OptionParser &parser) {
     parser.document_note(
-        "Implementation Notes",
+        "Implementation notes about the CEGAR algorithm",
         "The following describes differences of the implementation to "
         "the original implementation used and described in the paper.\n\n"
         "Conceptually, there is one larger difference which concerns the "
@@ -564,44 +766,24 @@ void add_implementation_notes_to_parser(options::OptionParser &parser) {
         "computation significantly while not having a strongly negative effect "
         "on heuristic quality due to potentially computing worse plans.\n\n"
         "Two further changes fix bugs of the original implementation to match "
-        "the description in the paper. The first concerns the single CEGAR "
-        "algorithm: raise a flaw for all goal variables of the task if the "
-        "plan for a PDB can be executed on the concrete task but does not lead "
-        "to a goal state. Previously, such flaws would not have been raised "
-        "because all goal variables are part of the collection from the start "
-        "on and therefore not considered. This means that the original "
-        "implementation accidentally disallowed merging patterns due to goal "
-        "violation flaws. The second bug fix is to actually randomize the "
-        "order of parallel operators in wildcard plan steps.",
+        "the description in the paper. The first bug fix is to raise a flaw "
+        "for all goal variables of the task if the plan for a PDB can be "
+        "executed on the concrete task but does not lead to a goal state. "
+        "Previously, such flaws would not have been raised because all goal "
+        "variables are part of the collection from the start on and therefore "
+        "not considered. This means that the original implementation "
+        "accidentally disallowed merging patterns due to goal violation "
+        "flaws. The second bug fix is to actually randomize the order of "
+        "parallel operators in wildcard plan steps.",
         true);
 }
 
-void add_cegar_options_to_parser(options::OptionParser &parser) {
-    parser.add_option<int>(
-        "max_pdb_size",
-        "maximum number of states per pattern database (ignored for the "
-        "initial collection consisting of singleton patterns for each goal "
-        "variable)",
-        "2000000",
-        Bounds("1", "infinity"));
-    parser.add_option<int>(
-        "max_collection_size",
-        "maximum number of states in the pattern collection (ignored for the "
-        "initial collection consisting of singleton patterns for each goal "
-        "variable)",
-        "20000000",
-        Bounds("1", "infinity"));
+void add_cegar_wildcard_option_to_parser(options::OptionParser &parser) {
     parser.add_option<bool>(
         "use_wildcard_plans",
         "if true, compute wildcard plans which are sequences of sets of "
         "operators that induce the same transition; otherwise compute regular "
         "plans which are sequences of single operators",
         "true");
-    parser.add_option<double>(
-        "max_time",
-        "maximum time in seconds for the CEGAR algorithm (ignored for"
-        "computing initial collection)",
-        "infinity",
-        Bounds("0.0", "infinity"));
 }
 }