+++ src/search/pdbs/pattern_collection_generator_disjoint_cegar.h
@@ -1,5 +1,5 @@
-#ifndef PDBS_PATTERN_COLLECTION_GENERATOR_SINGLE_CEGAR_H
-#define PDBS_PATTERN_COLLECTION_GENERATOR_SINGLE_CEGAR_H
+#ifndef PDBS_PATTERN_COLLECTION_GENERATOR_DISJOINT_CEGAR_H
+#define PDBS_PATTERN_COLLECTION_GENERATOR_DISJOINT_CEGAR_H
 
 #include "pattern_generator.h"
 
@@ -13,15 +13,15 @@ namespace pdbs {
   This pattern collection generator uses the CEGAR algorithm to compute a
   disjoint pattern collection for the given task. See cegar.h for more details.
 */
-class PatternCollectionGeneratorSingleCegar : public PatternCollectionGenerator {
+class PatternCollectionGeneratorDisjointCegar : public PatternCollectionGenerator {
     const int max_pdb_size;
     const int max_collection_size;
-    const bool use_wildcard_plans;
     const double max_time;
+    const bool use_wildcard_plans;
     const utils::Verbosity verbosity;
     std::shared_ptr<utils::RandomNumberGenerator> rng;
 public:
-    explicit PatternCollectionGeneratorSingleCegar(const options::Options &opts);
+    explicit PatternCollectionGeneratorDisjointCegar(const options::Options &opts);
 
     virtual PatternCollectionInformation generate(
         const std::shared_ptr<AbstractTask> &task) override;