+++ src/search/pdbs/utils.h
@@ -27,6 +27,19 @@ extern std::vector<FactPair> get_goals_in_random_order(
 extern std::vector<int> get_non_goal_variables(const TaskProxy &task_proxy);
 
 /*
+  Compute the causal graph neighbors for each variable of the task. If
+  bidirectional is false, then only predecessors of variables are considered
+  neighbors. If bidirectional is true, then the causal graph is treated as
+  undirected graph and also successors of variables are considered neighbors.
+*/
+extern std::vector<std::vector<int>> compute_cg_neighbors(
+    const std::shared_ptr<AbstractTask> &task,
+    bool bidirectional);
+
+extern PatternCollectionInformation get_pattern_collection_info(
+    const TaskProxy &task_proxy, const std::shared_ptr<PDBCollection> &pdbs);
+
+/*
   Dump the given pattern, the number of variables contained, the size of the
   corresponding PDB, and the runtime used for computing it. All output is
   prepended with the given string identifier.
@@ -45,6 +58,8 @@ extern void dump_pattern_collection_generation_statistics(
     const std::string &identifier,
     utils::Duration runtime,
     const PatternCollectionInformation &pci);
+
+extern std::string get_rovner_et_al_reference();
 }
 
 #endif